level2task2/src/main/java/ua/com/alevel/nix/pavlohnatenko/module1/Level2Task2.java
package ua.com.alevel.nix.pavlohnatenko.module1;


public class Level2Task2 {

    TreeNode root;

    static class TreeNode {
        int value;
        TreeNode left;
        TreeNode right;

        TreeNode(int value) {
            this.value = value;
            right = null;
            left = null;
        }

    }

    private TreeNode addRecursive(TreeNode current, int value) {
        if (current == null) {
            return new TreeNode(value);
        }

        if (value < current.value) {
            current.left = addRecursive(current.left, value);
        } else if (value > current.value) {
            current.right = addRecursive(current.right, value);
        } else {
            return current;
        }

        return current;
    }

    public void add(int value) {
        root = addRecursive(root, value);
    }

    public static int maxDepth(TreeNode node) {
        if (node == null)
            return 0;
        else {
            int lDepth = maxDepth(node.left);
            int rDepth = maxDepth(node.right);

            return Math.max(lDepth + 1, rDepth + 1);
        }
    }

}
